requirements.txt
requests
bs4
openai
python-telegram-bot==20.3
